refactor offer logic

+++ internal/offer/product.go
@@ -2,10 +2,10 @@ package offer
 
 type product struct {
 	name  string
-	price float32
+	price float64
 }
 
-func NewProduct(name string, price float32) (product, error) {
+func NewProduct(name string, price float64) (product, error) {
 	p := product{name, price}
 	if err := p.validate(); err != nil {
 		return product{}, err
@@ -17,6 +17,6 @@ func (p product) Name() string {
 	return p.name
 }
 
-func (p product) Price() float32 {
+func (p product) Price() float64 {
 	return p.price
 }
